Moved AtB attached entity generation out of personnel market.
Added an AtB event processor that can eventually be the location of the
bulk of the AtB code, on the way to pluginization.
